Improves date parsing
With this patch, it's now possible to parse those valid dates:

* 2014-11-10T23:38:51+0000
* 2014-11-10T23:38:51+00:00

Admittedly, the patch is rather messy. The idea comes from
http://stackoverflow.com/a/2202300/1796345 and the regexes have been used to
handle a few cases that didn't work out of the box.

I chose this messy code over using Joda time, since an advantage of aXMLRPC,
is that is has no dependencies

Fix #42